Fix some linux compile warnings

+++ mp/src/game/server/ff/ff_bot_temp.cpp
@@ -336,7 +336,7 @@ CON_COMMAND( bot_showhealth, "Makes a bot show his health" )
 
 		if( pPlayer && ( pPlayer->GetFlags() & FL_FAKECLIENT ) )
 		{
-			Warning( "[Bot %s] Health: %i (%i%%), Armor: %i (%i%%), Alive: %s: Lifestate: %i\n", pPlayer->GetPlayerName(), pPlayer->GetHealth(), pPlayer->GetHealthPercentage(), pPlayer->GetArmor(), pPlayer->GetArmorPercentage(), pPlayer->IsAlive() ? "Yes" : "No", pPlayer->m_lifeState );
+			Warning( "[Bot %s] Health: %i (%i%%), Armor: %i (%i%%), Alive: %s: Lifestate: %i\n", pPlayer->GetPlayerName(), pPlayer->GetHealth(), pPlayer->GetHealthPercentage(), pPlayer->GetArmor(), pPlayer->GetArmorPercentage(), pPlayer->IsAlive() ? "Yes" : "No", (int) pPlayer->m_lifeState );
 		}
 	}
 }
